+++ CHANGELOG.md
@@ -5,6 +5,19 @@
 
 ### Features
 
+* Capture the timestamp of when the message was enqueued on the source namespace in application properties. [Ben Dalling]
+
+* Allow the setting of ROUTER_TIMESTAMP_APP_PROPERTY_NAME. [Ben Dalling]
+
+### Build
+
+* Release/0.12.0. [Ben Dalling]
+
+
+## 0.11.0 (2025-09-23)
+
+### Features
+
 * Add the replay-dlq.py script. [Ben Dalling]
 
 ### Fix